@@ -5,28 +5,29 @@
     "start": "meteor run"
   },
   "dependencies": {
-    "@babel/runtime": "^7.0.0-beta.42",
+    "@babel/runtime": "^7.0.0-beta.44",
     "aws-sdk": "^2.224.1",
     "babel-runtime": "^6.26.0",
     "bcrypt": "^1.0.3",
-    "core-js": "^2.5.3",
-    "eslint-config-react-tools": "^1.0.10",
+    "core-js": "^2.5.5",
+    "eslint-config-react-tools": "^1.2.0",
     "eslint-plugin-jsx-a11y": "^6.0.3",
     "history": "^4.7.2",
-    "meteor-node-stubs": "~0.2.4",
+    "meteor-node-stubs": "^0.2.11",
     "normalize.css": "^8.0.0",
     "prop-types": "^15.6.1",
-    "react": "^16.2.0",
+    "react": "^16.3.1",
     "react-bootstrap": "^0.32.1",
+    "react-addons-css-transition-group": "^15.6.2",
     "react-bootstrap-sidebar": "0.0.1",
     "react-bootstrap-table": "^4.3.1",
-    "react-dom": "^16.2.0",
+    "react-dom": "^16.3.1",
     "react-icons-kit": "^1.0.7",
     "react-router": "^4.2.0",
     "react-router-dom": "^4.2.2",
-    "react-scripts": "^1.0.17",
+    "react-scripts": "^1.1.1",
     "react-sidebar": "^2.3.2",
     "react-sidenav": "^2.1.2",
-    "react-table": "^6.7.4"
+    "react-table": "^6.8.0"
   }
 }
